requirements.txt
Django==1.9
pyyaml
uwsgi
# REST API dependancies   
djangorestframework
markdown
django-filter
